requirements.txt
python-dotenv
numpy
pandas
scikit-learn
pyYAML
pymongo
certifi
mlflow
dagshub
fastapi
uvicorn
python-multipart
-e .